RAVE 0.434b
- Solucionat problema en el MOUSEWHEEL del DArbolEx i la DListaEx que no calculaba correctament el Item / Nodo resaltado.
- Afegit membre Longitud a les EtiquetaBD per sapiguer el tamany dels arxius de la etiqueta (FULL, s'ha corretgit la BD, les funcions de obtenir i guardar, i al ThreadAnalizar es calcula la longitud total.

+++ Reproductor de Audio y Video Extendido/ListaMedios.cpp
@@ -81,7 +81,7 @@ ItemMedio *ListaMedios::BuscarHash(sqlite3_int64 bHash) {
 void ListaMedios::Evento_MouseDobleClick(DWL::DEventoMouse &EventoMouse) {
 	if (EventoMouse.Boton == 0) {
 		if (_ItemResaltado != -1) {
-			MedioActual = _ItemResaltado;
+			MedioActual = static_cast<long>(_ItemResaltado);
 			ItemMedio *Itm = Medio(MedioActual);
 			
 			BDMedio NCan;
